+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.InputBox/Globals/GlobalDeclarations.cs
@@ -1,0 +1,20 @@
+﻿global using System;
+global using System.ComponentModel;
+global using System.ComponentModel.Design;
+global using System.Diagnostics;
+global using System.Drawing;
+global using System.Drawing.Design;
+global using System.Drawing.Drawing2D;
+global using System.Drawing.Imaging;
+global using System.IO;
+global using System.Reflection;
+global using System.Runtime.InteropServices;
+global using System.Security.Principal;
+global using System.Text;
+global using System.Windows.Forms;
+global using System.Windows.Forms.Design;
+global using System.Windows.Forms.VisualStyles;
+
+global using Krypton.Toolkit.Suite.Extended.InputBox.Properties;
+
+global using Microsoft.Win32;
