+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.InputBox/UX/KryptonInputBoxExtended.resx
